requirements.txt
pytorch
numpy
albumentations
wandb